Only one file for AC_CONFIG_SRCDIR. #4
configure: error: cannot find sources (cidrl.c cidrl6.c) in . or ..

— Macro: AC_CONFIG_SRCDIR (unique-file-in-source-dir)

unique-file-in-source-dir is some file that is in the package's source
directory; configure checks for this file's existence to make sure that
the directory that it is told contains the source code in fact does.
Occasionally people accidentally specify the wrong directory with
--srcdir; this is a safety check. See configure Invocation, for more
information.

https://www.gnu.org/software/autoconf/manual/autoconf-2.67/html_node/Input.html